Refactored zip and me accessors
