+++ accessors/zip/fixtures/TestReferenceCountNested.golden
@@ -1,102 +1,102 @@
 [
  {
-  "Base": "hello.txt",
+  "Base": "/hello.txt",
   "Data": "hello\n"
  },
  {
-  "Base": "hello1.txt",
+  "Base": "/hello1.txt",
   "Data": "hello1\n"
  },
  {
-  "Base": "hello2.txt",
+  "Base": "/hello2.txt",
   "Data": "hello2\n"
  },
  {
-  "Base": "hello3.txt",
+  "Base": "/hello3.txt",
   "Data": "hello3\n"
  },
  {
-  "Base": "hello4.txt",
+  "Base": "/hello4.txt",
   "Data": "hello4\n"
  },
  {
-  "Base": "hello.txt",
+  "Base": "/hello.txt",
   "Data": "hello\n"
  },
  {
-  "Base": "hello1.txt",
+  "Base": "/hello1.txt",
   "Data": "hello1\n"
  },
  {
-  "Base": "hello2.txt",
+  "Base": "/hello2.txt",
   "Data": "hello2\n"
  },
  {
-  "Base": "hello3.txt",
+  "Base": "/hello3.txt",
   "Data": "hello3\n"
  },
  {
-  "Base": "hello4.txt",
+  "Base": "/hello4.txt",
   "Data": "hello4\n"
  },
  {
-  "Base": "hello.txt",
+  "Base": "/hello.txt",
   "Data": "hello\n"
  },
  {
-  "Base": "hello1.txt",
+  "Base": "/hello1.txt",
   "Data": "hello1\n"
  },
  {
-  "Base": "hello2.txt",
+  "Base": "/hello2.txt",
   "Data": "hello2\n"
  },
  {
-  "Base": "hello3.txt",
+  "Base": "/hello3.txt",
   "Data": "hello3\n"
  },
  {
-  "Base": "hello4.txt",
+  "Base": "/hello4.txt",
   "Data": "hello4\n"
  },
  {
-  "Base": "hello.txt",
+  "Base": "/hello.txt",
   "Data": "hello\n"
  },
  {
-  "Base": "hello1.txt",
+  "Base": "/hello1.txt",
   "Data": "hello1\n"
  },
  {
-  "Base": "hello2.txt",
+  "Base": "/hello2.txt",
   "Data": "hello2\n"
  },
  {
-  "Base": "hello3.txt",
+  "Base": "/hello3.txt",
   "Data": "hello3\n"
  },
  {
-  "Base": "hello4.txt",
+  "Base": "/hello4.txt",
   "Data": "hello4\n"
  },
  {
-  "Base": "hello.txt",
+  "Base": "/hello.txt",
   "Data": "hello\n"
  },
  {
-  "Base": "hello1.txt",
+  "Base": "/hello1.txt",
   "Data": "hello1\n"
  },
  {
-  "Base": "hello2.txt",
+  "Base": "/hello2.txt",
   "Data": "hello2\n"
  },
  {
-  "Base": "hello3.txt",
+  "Base": "/hello3.txt",
   "Data": "hello3\n"
  },
  {
-  "Base": "hello4.txt",
+  "Base": "/hello4.txt",
   "Data": "hello4\n"
  }
 ]
