--- vql/filesystem/zip_windows.go
@@ -1,6 +1,0 @@
-package filesystem
-
-import (
-	// Load windows file accessor for tests
-	_ "www.velocidex.com/golang/velociraptor/vql/windows/filesystems"
-)
